+++ CHANGELOG.md
@@ -53,6 +53,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Upgrade to Spring Boot 3
     - This affects the capability of the path matcher
 
+- Added option to calculate checksum for each import file ([#1015](https://github.com/adorsys/keycloak-config-cli/issues/1015))
+
 ## [5.12.0] - 2024-03-28
 - Added support for managing message bundles